feat: conversión a componentes el update
